Handle more than 50 missing game licenses
