+++ types/types.d.ts
@@ -41,6 +41,10 @@ declare class Bot {
      */
     checkMsgBlock(steamID64: any, message: string): boolean;
     /**
+     * Handles checking for missing game licenses, requests them and then starts playing
+     */
+    handleMissingGameLicenses(): void;
+    /**
      * Our commandHandler respondModule implementation - Sends a message to a Steam user
      * @param _this - The Bot object context
      * @param resInfo - Object containing information passed to command by friendMessage event
@@ -99,6 +103,10 @@ declare class Bot {
      * Handles aborting a login attempt should an account get stuck to prevent the bot from softlocking (see issue #139)
      */
     handleLoginTimeout(): void;
+    /**
+     * Handles checking for missing game licenses, requests them and then starts playing
+     */
+    handleMissingGameLicenses(): void;
     /**
      * Our commandHandler respondModule implementation - Sends a message to a Steam user
      * @param _this - The Bot object context
